switch VectorType to migration_types
