+++ store/backend/neurostore/models/migration_types.py
@@ -1,5 +1,6 @@
 import sqlalchemy as sa
 from sqlalchemy.dialects.postgresql import TSVECTOR
+from pgvector.sqlalchemy import Vector
 
 
 class TSVector(sa.types.TypeDecorator):
@@ -7,3 +8,11 @@ class TSVector(sa.types.TypeDecorator):
 
     cache_ok = True
     impl = TSVECTOR
+
+
+class VectorType(sa.types.TypeDecorator):
+    cache_ok = True
+    impl = Vector
+
+    def load_dialect_impl(self, dialect):
+        return dialect.type_descriptor(Vector())
